Cast boolean import to integer so it's valid in SQL query. Closes #673

+++ modules/pm/pm.useredit.first.php
@@ -17,6 +17,6 @@ Hooks=users.profile.update.first,users.edit.update.first
 
 defined('COT_CODE') or die('Wrong URL.');
 
-$ruser['user_pmnotify'] = cot_import('ruserpmnotify','P','BOL');
+$ruser['user_pmnotify'] = (int)cot_import('ruserpmnotify','P','BOL');
 
 ?>
